requirements.txt
Flask==3.1.0
pika==1.3.2
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.3